requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
decompyle3==3.9.2
defusedxml==0.8.0rc2
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
idna==3.10
mysqlclient==2.2.6
oauthlib==3.2.2
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
requests==2.26.0
requests-oauthlib==2.0.0
semantic-version==2.10.0
setuptools-rust==1.10.2
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
spark-parser==1.8.9
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uncompyle6==3.9.2
urllib3==1.26.20
xdis==6.1.1
